requirements.txt
streamlit>=1.28.0
streamlit-option-menu>=0.3.6


# Database & Spatial Data
psycopg2-binary>=2.9.7
SQLAlchemy>=2.0.0
geopandas>=0.14.0
shapely>=2.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0


# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0


plotly>=5.15.0
folium>=0.14.0
streamlit-folium>=0.13.0
altair>=5.0.0
matplotlib>=3.7.0
seaborn>=0.12.0


# Utility & Performance
requests>=2.31.0
cachetools>=5.3.0
retry>=0.9.2